feat: extract Overpass API queries to a data loader
